mondif integer en float

+++ db/migrate/20220404193529_create_litrages.rb
@@ -2,9 +2,9 @@ class CreateLitrages < ActiveRecord::Migration[6.1]
   def change
     create_table :litrages do |t|
       t.integer :kg
-      t.integer :debutant
-      t.integer :intermediaire
-      t.integer :confirme
+      t.float :debutant
+      t.float :intermediaire
+      t.float :confirme
 
       t.timestamps
     end
